Add quick record mode
Some of application implemented auto memory manage logic, which will
encounter error if the memory layout was changed by traced_malloc.
so, the quick record mode was added, which will not change the memory
layout, but will just record each malloc and free memory, and one
offline tools will be needed to analysis the leaked memory.